requirements.txt
gunicorn
flask == 2.3.1
google-generativeai
replicate
